requirements.txt
torch==1.9.0
onnx== 1.9.0
torchvision==0.10.1 # Image classification
onnx-simplifier==0.4.8
onnxruntime==1.8.0
opencv-python==4.6.0